MT#55283 optimise timer tree lookup
Look up the first object in the tree to determine the needed sleep time.
But instead of leaving the object in the tree for the next (post-sleep)
iteration, which requires another tree lookup, remove it from the tree
and remember it, anticipating that nothing would get added into the tree
to be scheduled sooner. This saves us from having to do another tree
lookup for each timer that runs.

Adapt the scheduling function to be able to handle this and wake up the
thread sooner if necessary.

Change-Id: I9297346bc4aa8d2f68ecb833be8f71ce62a3bbfe